Check unit of ps rather than unit of ap
Check unit of ps rather than unit of ap: the unit of ps is more likely
to be present. Do this before reading ap and b. So we have to keep the
NetCDF ID of ps in a new variable `varid_p`. `varid_p` is also used for
the 3D pressure if not hybrid. This is clearer anyway than keeping the
ID of ps or 3D pressure in variable varid.