Open a separate pressure file in Fortran program
If the pressure file is specified on the command line of `ml2pl.sh`,
open it from the Fortran program and look for pressure variables in
that pressure file, instead of creating a unique input file with
ncks. The advantage is that we spare two copying processes (cp and
ncks), which can take several minutes with large files.